Revert renaming of solver module
