+++ pymanopt/solvers/conjugate_gradient.py
@@ -4,8 +4,8 @@ from copy import deepcopy
 
 import numpy as np
 
-from pymanopt.solvers._solver import Solver
 from pymanopt.solvers.line_search import AdaptiveLineSearcher
+from pymanopt.solvers.solver import Solver
 from pymanopt.tools import printer
