+++ pymanopt/solvers/solver.py
